requirements.txt
flask
pymongo
python-dotenv
gunicorn
flask-wtf
flask-security
passlib
email_validator
